[GOBBLIN-1693] Move kafka ingestion health model logic to separate class
Moving health modeling logic to a separate class to avoid clutter when adding
helix dynamic work unit allocation feature. The plan is to send a helix dynamic
work unit split message when this health check fails.

To comply with single responsibility principle, the KafkaIngestionHealthCheck will
act more as a controller and the actual health model will be in a new KafkaHealthModel
class for business logic.

Java doc and GTE info